backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
packaging==21.3
PyMySQL==1.0.2
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
wincertstore==0.2
